+++ includes/general.js
@@ -1,8 +1,8 @@
 var month_names_short = ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'];
 
-exports.getDate = getDate;
+// Function to Style the date
 function getDate(post) {
-  var date = post.getDate('post.date');
+  var date = new Date(post.data.date);
   if (date) {
     var suffix = '';
     switch (date.getUTCDate().toString().substr(date.getUTCDate().toString().length - 1)) {
@@ -19,10 +19,38 @@ function getDate(post) {
       suffix = 'th';
     }
 
-    var formatedDate = month_names_short[date.getUTCMonth()] + ', ' +
-            date.getUTCDate() + suffix + ' ' + date.getUTCFullYear();
+    var formatedDate = month_names_short[date.getUTCMonth()] + ' ' +
+            date.getUTCDate() + suffix + ', ' + date.getUTCFullYear();
     return formatedDate;
   } else {
     return '';
   }
+}
+
+// Function to get the first paragraph of a post
+function getFirstParagraph(post) {
+  var slices = post.data.body;
+  var firstParagraph = "";
+  var haveFirstParagraph = false;
+  
+  slices.forEach(function(slice) {
+    if (!haveFirstParagraph) {
+      if (slice.slice_type == "text") {
+        slice.primary.text.forEach(function(block){
+          if (block.type == "paragraph") {
+            if (!haveFirstParagraph) {
+              firstParagraph += block.text;
+              haveFirstParagraph = true;
+            }
+          }
+        });
+      }
+    }
+  });
+  return firstParagraph;
+}
+
+module.exports = {
+  getDate: getDate,
+  getFirstParagraph: getFirstParagraph,
 }
